Cleaned helper function and fixed duplicate run checks
Cleaned helper function and fixed duplicate run checks
